auto test generate mono-repo

+++ test/auto-test/files/3_web_gin_pb.sh
@@ -3,50 +3,87 @@
 webServiceName="user"
 webDir="3_web_gin_pb_${webServiceName}"
 
+colorCyan='\e[1;36m'
+markEnd='\e[0m'
+errCount=0
+
 function checkResult() {
-    result=$1
-    if [ ${result} -ne 0 ]; then
-        exit ${result}
-    fi
+  result=$1
+  if [ ${result} -ne 0 ]; then
+      exit ${result}
+  fi
+}
+
+function checkErrCount() {
+  result=$1
+  if [ ${result} -ne 0 ]; then
+      ((errCount++))
+  fi
 }
 
 function stopService() {
-  pid=$(ps -ef | grep "./cmd/${webServiceName}" | grep -v grep | awk '{print $2}')
+  local name=$1
+  if [ "$name" == "" ]; then
+    echo "name cannot be empty"
+    exit 1
+  fi
+
+  local processMark="./cmd/$name"
+  pid=$(ps -ef | grep "${processMark}" | grep -v grep | awk '{print $2}')
   if [ "${pid}" != "" ]; then
       kill -9 ${pid}
   fi
 }
 
+function checkServiceStarted() {
+  local name=$1
+  if [ "$name" == "" ]; then
+    echo "name cannot be empty"
+    exit 1
+  fi
+
+  local processMark="./cmd/$name"
+  local timeCount=0
+  # waiting for service to start
+  while true; do
+    sleep 1
+    pid=$(ps -ef | grep "${processMark}" | grep -v grep | awk '{print $2}')
+    if [ "${pid}" != "" ]; then
+        break
+    fi
+    (( timeCount++ ))
+    if (( timeCount >= 30 )); then
+      echo "service startup timeout"
+      exit 1
+    fi
+  done
+}
+
 function testRequest() {
-  echo "--------------------- 20s 后测试开始 ---------------------"
-  sleep 20
+  checkServiceStarted $webServiceName
+  sleep 1
+  echo "--------------------- start testing ---------------------"
 
   echo -e "\n\n"
-  echo 'curl -X POST http://localhost:8080/api/v1/auth/register -H "Content-Type: application/json" -d "{\"email\":\"foo@bar.com\",\"password\":\"123456\"}"'
+  echo -e "${colorCyan}curl -X POST http://localhost:8080/api/v1/auth/register -H \"Content-Type: application/json\" -d {\"email\":\"foo@bar.com\",\"password\":\"123456\"} ${markEnd}"
   curl -X POST http://localhost:8080/api/v1/auth/register -H "Content-Type: application/json" -d "{\"email\":\"foo@bar.com\",\"password\":\"123456\"}"
-  echo -e "\n\n"
-  sleep 3
+  checkErrCount $?
 
   echo -e "\n\n"
-  echo 'curl -X POST http://localhost:8080/api/v1/auth/register  -H "Content-Type: application/json" -H "X-Request-Id: qaz12wx3ed4" -d "{\"email\":\"foo@bar.com\",\"password\":\"123456\"}"'
+  echo -e "${colorCyan}curl -X POST http://localhost:8080/api/v1/auth/register  -H "Content-Type: application/json" -H \"X-Request-Id: qaz12wx3ed4\" -d {\"email\":\"foo@bar.com\",\"password\":\"123456\"} ${markEnd}"
   curl -X POST http://localhost:8080/api/v1/auth/register -H "Content-Type: application/json" -H "X-Request-Id: qaz12wx3ed4" -d "{\"email\":\"foo@bar.com\",\"password\":\"123456\"}"
-  echo -e "\n\n"
-  sleep 3
+  checkErrCount $?
 
-  echo "--------------------- 测试结束！---------------------"
-  stopService
+  echo -e "\n--------------------- the test is over, error result: $errCount ---------------------\n"
+  stopService $webServiceName
 }
 
 if [ -d "${webDir}" ]; then
-  echo "web服务 ${webDir} 已存在"
+  echo "service ${webDir} already exists"
 else
-  echo "创建web服务 ${webDir}"
-  sponge web http-pb \
-    --module-name=${webServiceName} \
-    --server-name=${webServiceName} \
-    --project-name=ginpbdemo \
-    --protobuf-file=./files/user.proto \
-    --out=./${webDir}
+  echo "create service ${webDir}"
+  echo -e "${colorCyan}sponge web http-pb --module-name=${webServiceName} --server-name=${webServiceName} --project-name=ginpbdemo --protobuf-file=./files/user.proto --out=./${webDir} ${markEnd}"
+  sponge web http-pb --module-name=${webServiceName} --server-name=${webServiceName} --project-name=ginpbdemo --protobuf-file=./files/user.proto --out=./${webDir}
   checkResult $?
 fi
 
@@ -57,11 +94,10 @@ echo "make proto"
 make proto
 checkResult $?
 
-echo "替换示例模板代码"
+echo "replace the sample template code"
 replaceCode ../files/web_gin_pb_content ./internal/handler/user.go
 checkResult $?
 
-echo "test request"
 testRequest &
 
 echo "make run"
