+++ test/auto-test/files/4_micro_grpc_pb.sh
@@ -3,45 +3,85 @@
 grpcServiceName="user"
 grpcDir="4_micro_grpc_pb_${grpcServiceName}"
 
+colorCyan='\e[1;36m'
+markEnd='\e[0m'
+errCount=0
+
 function checkResult() {
-    result=$1
-    if [ ${result} -ne 0 ]; then
-        exit ${result}
-    fi
+  result=$1
+  if [ ${result} -ne 0 ]; then
+      exit ${result}
+  fi
+}
+
+function checkErrCount() {
+  result=$1
+  if [ ${result} -ne 0 ]; then
+      ((errCount++))
+  fi
 }
 
 function stopService() {
-  pid=$(ps -ef | grep "./cmd/${grpcServiceName}" | grep -v grep | awk '{print $2}')
+  local name=$1
+  if [ "$name" == "" ]; then
+    echo "name cannot be empty"
+    exit 1
+  fi
+
+  local processMark="./cmd/$name"
+  pid=$(ps -ef | grep "${processMark}" | grep -v grep | awk '{print $2}')
   if [ "${pid}" != "" ]; then
       kill -9 ${pid}
   fi
 }
 
-function testRequest() {
-  echo "--------------------- 20s 后测试开始 ---------------------"
-  sleep 20
-  cd internal/service
-  for i in {1..3}; do
-    echo -e "\n\n"
-    echo "${i} go test -run Test_service_user_methods/Register"
-    go test -run Test_service_user_methods/Register
-    echo -e "\n\n"
+function checkServiceStarted() {
+  local name=$1
+  if [ "$name" == "" ]; then
+    echo "name cannot be empty"
+    exit 1
+  fi
+
+  local processMark="./cmd/$name"
+  local timeCount=0
+  # waiting for service to start
+  while true; do
+    sleep 1
+    pid=$(ps -ef | grep "${processMark}" | grep -v grep | awk '{print $2}')
+    if [ "${pid}" != "" ]; then
+        break
+    fi
+    (( timeCount++ ))
+    if (( timeCount >= 30 )); then
+      echo "service startup timeout"
+      exit 1
+    fi
   done
+}
+
+function testRequest() {
+  checkServiceStarted $grpcServiceName
+  sleep 1
+  echo "--------------------- start testing ---------------------"
+
+  cd internal/service
+
+  echo -e "\n\n"
+  echo -e "${colorCyan}go test -run Test_service_user_methods/Register ${markEnd}"
+  go test -run Test_service_user_methods/Register
+  checkErrCount $?
+
   cd -
-  echo "--------------------- 测试结束！---------------------"
-  stopService
+  echo -e "\n--------------------- the test is over, error result: $errCount ---------------------\n"
+  stopService $grpcServiceName
 }
 
 if [ -d "${grpcDir}" ]; then
-  echo "微服务 ${grpcDir} 已存在"
+  echo "service ${grpcDir} already exists"
 else
-  echo "创建微服务 ${grpcDir}"
-  sponge micro rpc-pb \
-    --module-name=${grpcServiceName} \
-    --server-name=${grpcServiceName} \
-    --project-name=grpcpbdemo \
-    --protobuf-file=./files/user2.proto \
-    --out=./${grpcDir}
+  echo "create service ${grpcDir}"
+  echo -e "${colorCyan}sponge micro rpc-pb --module-name=${grpcServiceName} --server-name=${grpcServiceName} --project-name=grpcpbdemo --protobuf-file=./files/user2.proto --out=./${grpcDir} ${markEnd}"
+  sponge micro rpc-pb --module-name=${grpcServiceName} --server-name=${grpcServiceName} --project-name=grpcpbdemo --protobuf-file=./files/user2.proto --out=./${grpcDir}
   checkResult $?
 fi
 
@@ -52,13 +92,12 @@ echo "make proto"
 make proto
 checkResult $?
 
-echo "替换示例模板代码"
+echo "replace the sample template code"
 replaceCode ../files/micro_grpc_pb_content ./internal/service/user2.go
 checkResult $?
-replaceCode ../files/micro_grpc_pb_content2 ./internal/service/user2_client_test.go
+replaceCode ../files/micro_grpc_pb_content ./internal/service/user2_client_test.go
 checkResult $?
 
-echo "test request"
 testRequest &
 
 echo "make run"
